Rewrote shard-dist test.
